+++ package/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 5.0.1
+
+- Updated documentation
+
+## 5.0.0
+
+- Refactoring 'ModelRepository'
+- Refactoring 'ModelTTNamedTTNamedTTNamedTTIterator'
+- Updated documentation
+
 ## 4.0.0
 
 - A new 'TempCacheProvider' object has been created, with which you can share data between 'NamedVM' directly. The reason for this was a data leak problem in 'TempCacheService', it was impossible to close a stream in a certain 'NamedVM'. And there was also a problem in that 'TempCacheService' had to connect to 'ModelRepository' in order to then get the data we needed from 'NamedVM' via 'ModelRepository'. Of course, this took a lot of time, but now it doesn't